+++ CHANGELOG.md
@@ -11,8 +11,8 @@ and this project adheres to
 ### Changed
 
 - Use anybuf from crates.io
-- drand: Add `QueryMsg::IsAllowlisted` (one word) analogue to
-  `QueryMsg::IsAllowListed`.
+- drand: Add `QueryMsg::IsAllowlisted` and `QueryMsg::Allowlist` (one word)
+  analogue to `QueryMsg::IsAllowListed` and `QueryMsg::AllowList`.
 - proxy: Add allowlist
 
 ## [0.13.2] - 2023-04-26